+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 View full changelog [here](https://github.com/shedaniel/RoughlyEnoughItems/blob/1.14/CHANGELOG.md).
+## v2.9-beta+build.106 (BETA)
+- Using: [HammerLib](https://minecraft.curseforge.com/projects/hammer-lib) as a simple opengl scissors api
+- New: Mod Name of category new shows in category tooltips
+- New: Renderer API
+- New: Villager Trading Like Recipe Screen. ![image](https://cdn.discordapp.com/attachments/432055962233470988/576077676305973275/unknown.png)
+- New: Recipe Screen Selection Screen
+- Removed: All `Locale.ROOT` usage
+- Fixed: Button Width Rendering issues when the width is an odd number
+- Changed: Mod Name changed from `RoughlyEnoughItems` to `Roughly Enough Items`
+- Changed: Lots of internal refractors
 ## v2.8.2+build.104
 - Fixed [#81](https://github.com/shedaniel/RoughlyEnoughItems/issues/81): Scrolling unaffected by exclusion zones.
 - Added [#82](https://github.com/shedaniel/RoughlyEnoughItems/issues/82): Close search after pressing "Enter"